Moved the class and functions for password validation in a separate source file to organize the frontend.
